requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
pyyaml==6.0.2
tqdm==4.66.4
Pillow==10.4.0
tabulate==0.9.0
rich==13.7.1
PyMuPDF==1.26.3
pdfminer.six==20250506

python-dotenv==1.0.1
# Prefer the new Google GenAI SDK (images, video). If unavailable in your env, you can switch to google-generativeai.
google-genai>=0.3.0
google-generativeai>=0.7.0

